modify max latency

+++ services/data_process/src/pipeline_node/multimedia_codec/decoder/decode_data_process.cpp
@@ -170,7 +170,7 @@ int32_t DecodeDataProcess::InitDecoderMetadataFormat()
     metadataFormat_.PutStringValue("codec_mime", processType_);
     metadataFormat_.PutIntValue("width", sourceConfig_.GetWidth());
     metadataFormat_.PutIntValue("height", sourceConfig_.GetHeight());
-    metadataFormat_.PutIntValue("frame_rate", MAX_FRAME_RATE);
+    metadataFormat_.PutDoubleValue("frame_rate", MAX_FRAME_RATE);
 
     return DCAMERA_OK;
 }
